fix(backups): restore operator-managed storage access
